Pass document as input to paint timing algorithms
Since these algorithms are called from the event loop directly, we need to specify the Documents explicitly. In addition, https://github.com/w3c/performance-timeline/issues/162 has been filed to track the change to Performance Timeline to support receiving the global object as argument to the "queue an entry" algorithm.