Add `PaintTimingMixin`, which exposes two timestamps: - `paintTime` (the end of "update the rendering phase") - `presentationTime` (an optional implementation-defined timestamp)
`startTime` returns `presentationTime`, or `paintTime` if `presentationTime`
is not implemented.

Currently this `PaintTimingMixin` is only included in `PerformancePaintTiming`,
but it will be included in `ElementTiming`, `LargestContentfulPaint`,
`PerformanceEventTiming`, and `PerformanceLongAnimationFrameTiming` as a follow up.

This is based on a resolution at the web performance working group.
Minutes: https://docs.google.com/document/d/1ZWAUJZBJUvSUyShvKXNEU-cuCe47jpgbR69ckWZUTfI/edit?tab=t.0#heading=h.kb3idfbysfg7

Closes #62